Add samples for some of the functions
