+++ src/stream/mod.rs
@@ -307,8 +307,9 @@ pub use once::{once, Once};
 pub use repeat::{repeat, Repeat};
 pub use repeat_with::{repeat_with, RepeatWith};
 pub use stream::*;
+pub use crate::stream::sample::Sample;
 
-pub(crate) mod stream;
+pub mod stream;
 
 mod empty;
 mod from_fn;
@@ -316,6 +317,7 @@ mod from_iter;
 mod once;
 mod repeat;
 mod repeat_with;
+mod sample;
 
 cfg_unstable! {
     pub mod double_ended;
